trying to better the bounding

+++ src/classes/Pallet.cpp
@@ -1,7 +1,7 @@
 #include "Pallet.h"
 
 Pallet::Pallet(int id,int weight, int profit) : id(id),
-           weight(weight), profit(profit), ratio(static_cast<float>(profit) / weight) {}
+           weight(weight), profit(profit), ratio(static_cast<double>(profit) / weight) {}
 
 int Pallet::getId()    const{
   return id;
@@ -15,6 +15,6 @@ int Pallet::getProfit() const {
     return profit;
 }
 
-int Pallet::getRatio() const {
+double Pallet::getRatio() const {
   return ratio;
 }
